:art: :racehorse: reduce duplicate command object instances by lazy-loading them
