+++ lib/cli.js
@@ -20,7 +20,7 @@ function cli(args) {
             }
 
             // output help information for the particular command
-            cmdInstances[command].commander.help();
+            cmdInstances[command].help();
         });
 
     program.option('-V, --verbose', 'Enable verbose output');
