+++ lib/commands/buster.js
@@ -1,9 +1,11 @@
 var BaseCommand = require('./base');
 
-module.exports = BaseCommand.extend({
+module.exports = {
     name: 'buster',
-    description: 'who ya gonna call?',
+    description: 'who ya gonna call?'
+};
 
+module.exports.Command = BaseCommand.extend({
     execute: function () {
         var npm = require('../utils/npm');
