+++ lib/commands/config/index.js
@@ -2,7 +2,7 @@ var BaseCommand = require('../base'),
     advancedOptions = require('./advanced'),
     validator = require('validator');
 
-module.exports = BaseCommand.extend({
+module.exports = {
     name: 'config',
     description: 'configure a Ghost instance',
 
@@ -19,8 +19,10 @@ module.exports = BaseCommand.extend({
         alias: 'e',
         description: 'Environment to get/set config for',
         defaultValue: 'production'
-    }],
+    }].concat(advancedOptions)
+};
 
+module.exports.Command = BaseCommand.extend({
     prompts: [{
         type: 'input',
         name: 'url',
@@ -31,12 +33,6 @@ module.exports = BaseCommand.extend({
             }) || 'Invalid url!';
         }
     }],
-
-    init: function () {
-        this.options = this.options.concat(advancedOptions);
-
-        this._super.apply(this, arguments);
-    },
 
     execute: function (key, value, options) {
         // ensure we are within a valid Ghost install
