+++ lib/commands/doctor/index.js
@@ -1,13 +1,15 @@
 var BaseCommand = require('../base');
 
-module.exports = BaseCommand.extend({
+module.exports = {
     name: 'doctor',
     description: 'check the system for any potential hiccups when installing/updating Ghost',
 
     arguments: [
         {name: 'category', optional: true}
-    ],
+    ]
+};
 
+module.exports.Command = BaseCommand.extend({
     categories: {
         install: {
             nodeVersion: 'System is running supported version of node',
