+++ lib/commands/install.js
@@ -1,6 +1,7 @@
-var BaseCommand = require('./base');
+var BaseCommand = require('./base'),
+    advancedOptions = require('./config/advanced');
 
-module.exports = BaseCommand.extend({
+module.exports = {
     name: 'install',
     description: 'install a brand new instance of Ghost',
 
@@ -16,15 +17,10 @@ module.exports = BaseCommand.extend({
         alias: 'N',
         description: 'Don\'t automatically run the setup command',
         flag: true
-    }],
+    }].concat(advancedOptions)
+};
 
-    init: function () {
-        var advancedOptions = require('./config/advanced');
-
-        this.options = this.options.concat(advancedOptions);
-        this._super.apply(this, arguments);
-    },
-
+module.exports.Command = BaseCommand.extend({
     execute: function (version, options) {
         var Promise         = require('bluebird'),
             path            = require('path'),
