+++ lib/commands/setup.js
@@ -1,6 +1,7 @@
-var BaseCommand = require('./base');
+var BaseCommand = require('./base'),
+    advancedOptions = require('./config/advanced');
 
-module.exports = BaseCommand.extend({
+module.exports = {
     name: 'setup',
     description: 'setup an installation of Ghost (after it is installed)',
 
@@ -9,20 +10,10 @@ module.exports = BaseCommand.extend({
         alias: 'l',
         description: 'Quick setup for a local installation of Ghost',
         flag: true
-    }],
+    }].concat(advancedOptions)
+};
 
-    init: function (program) {
-        // Command is being run as a top-level command, go ahead and mount
-        // the advanced options
-        if (program) {
-            var advancedOptions = require('./config/advanced');
-
-            this.options = this.options.concat(advancedOptions);
-        }
-
-        this._super.apply(this, arguments);
-    },
-
+module.exports.Command = BaseCommand.extend({
     execute: function (options) {
         this.checkValidInstall();
