+++ lib/commands/start.js
@@ -1,6 +1,6 @@
 var BaseCommand = require('./base');
 
-module.exports = BaseCommand.extend({
+module.exports = {
     name: 'start',
     description: 'starts an instance of Ghost',
 
@@ -14,11 +14,10 @@ module.exports = BaseCommand.extend({
         alias: 'P',
         description: 'Start ghost in production mode',
         flag: true
-    }],
+    }]
+};
 
-    // TODO: this currently only supports pm2 process running.
-    // This should be extended to support custom methods of stopping/starting
-    // ghost
+module.exports.Command = BaseCommand.extend({
     execute: function (options) {
         // ensure we are within a valid ghost install
         this.checkValidInstall();
