+++ lib/commands/stop.js
@@ -1,12 +1,11 @@
 var BaseCommand = require('./base');
 
-module.exports = BaseCommand.extend({
+module.exports = {
     name: 'stop',
-    description: 'stops a named instance of Ghost',
+    description: 'stops a named instance of Ghost'
+};
 
-    // TODO: this currently only supports pm2 process running.
-    // This should be extended to support custom methods of stopping/starting
-    // ghost
+module.exports.Command = BaseCommand.extend({
     execute: function () {
         // ensure we are within a valid ghost install
         this.checkValidInstall();
