+++ lib/commands/update.js
@@ -1,6 +1,6 @@
 var BaseCommand = require('./base');
 
-module.exports = BaseCommand.extend({
+module.exports = {
     name: 'update',
     description: 'update a Ghost instance',
 
@@ -17,8 +17,10 @@ module.exports = BaseCommand.extend({
         alias: 'f',
         description: 'Force Ghost to update',
         flag: true
-    }],
+    }]
+};
 
+module.exports.Command = BaseCommand.extend({
     execute: function (version, options) {
         this.checkValidInstall();
